ETHBE-734: Use common approach for '/v1/.../eth_transfers' pagination tests
